BC-662 | Fixes to packages
