+++ laa-amend-a-claim-ui/src/main/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/AmendmentDateValidator.java
@@ -1,9 +1,10 @@
 package uk.gov.justice.laa.amend.claim.forms.amendments.validators;
 
-import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
 import java.time.YearMonth;
 import java.time.format.DateTimeFormatter;
+import java.time.format.DateTimeFormatterBuilder;
+import java.util.Locale;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
 import org.springframework.validation.Errors;
@@ -20,8 +21,15 @@ public abstract class AmendmentDateValidator implements FieldSpecificAmendmentVa
       "amendmentForm.dates.cantBeLaterThanSubmissionPeriod";
 
   private static final String DATE_FORMAT_D_MMM_YYYY = "d MMMM yyyy";
-  protected static final DateTimeFormatter DATE_FORMATTER_D_MMM_YYYY =
+  private static final DateTimeFormatter DATE_FORMATTER_D_MMM_YYYY =
       DateTimeFormatter.ofPattern(DATE_FORMAT_D_MMM_YYYY);
+
+  private static final String DATE_FORMAT_MMM_YYYY = "MMM-yyyy";
+  public static final DateTimeFormatter SUBMISSION_PERIOD_FORMATTER =
+      new DateTimeFormatterBuilder()
+          .parseCaseInsensitive()
+          .appendPattern(DATE_FORMAT_MMM_YYYY)
+          .toFormatter(Locale.ENGLISH);
 
   protected void addDateTooEarlyMessage(
       Errors errors, ClaimViewField<?> field, LocalDate earliestDate) {
@@ -45,7 +53,7 @@ public abstract class AmendmentDateValidator implements FieldSpecificAmendmentVa
         errors);
   }
 
-  protected final LocalDate getSubmissionPeriodCutoffDate(@NotNull YearMonth submissionPeriod) {
+  protected LocalDate getTwentiethOfNextMonth(YearMonth submissionPeriod) {
     return submissionPeriod.plusMonths(1).atDay(20);
   }
 }
