+++ laa-amend-a-claim-ui/src/main/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/CaseStartDateValidator.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForm;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
-import uk.gov.justice.laa.amend.claim.utils.DateWrapperUtil;
 import uk.gov.justice.laa.amend.claim.viewmodels.viewfield.ClaimDetailsViewField;
 import uk.gov.justice.laa.amend.claim.viewmodels.viewfield.ClaimViewField;
 
@@ -16,11 +15,8 @@ public class CaseStartDateValidator extends AmendmentDateValidator {
 
   private static final LocalDate OLDEST_DATE_ALLOWED = LocalDate.of(1995, Month.JANUARY, 1);
 
-  private final DateWrapperUtil dateWrapperUtil;
-
-  public CaseStartDateValidator(MessageSource messageSource, DateWrapperUtil dateWrapperUtil) {
+  public CaseStartDateValidator(MessageSource messageSource) {
     super(messageSource);
-    this.dateWrapperUtil = dateWrapperUtil;
   }
 
   @Override
@@ -31,16 +27,10 @@ public class CaseStartDateValidator extends AmendmentDateValidator {
   @Override
   public void validate(
       ClaimDetails claim, ClaimViewField<?> field, AmendmentForm form, Errors errors) {
-    var caseStartDate = form.getDateValue(field.name());
+    var caseStartDate = form.getDateValue(ClaimDetailsViewField.CASE_START_DATE.toString());
 
-    if (caseStartDate == null) {
-      return;
-    }
-
-    if (dateWrapperUtil.isFutureDate(caseStartDate)) {
-      addDateInTheFutureMessage(errors, field);
-    } else if (caseStartDate.isBefore(OLDEST_DATE_ALLOWED)) {
-      addDateTooEarlyMessage(errors, field, OLDEST_DATE_ALLOWED);
+    if (caseStartDate != null && caseStartDate.isBefore(OLDEST_DATE_ALLOWED)) {
+      addDateTooEarlyMessage(errors, ClaimDetailsViewField.CASE_START_DATE, OLDEST_DATE_ALLOWED);
     }
   }
 }
