+++ laa-amend-a-claim-ui/src/main/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/CivilTransferDateValidator.java
@@ -7,7 +7,6 @@ import org.springframework.stereotype.Component;
 import org.springframework.validation.Errors;
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForm;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
-import uk.gov.justice.laa.amend.claim.utils.DateWrapperUtil;
 import uk.gov.justice.laa.amend.claim.viewmodels.viewfield.CivilClaimDetailsViewField;
 import uk.gov.justice.laa.amend.claim.viewmodels.viewfield.ClaimViewField;
 
@@ -16,11 +15,8 @@ public class CivilTransferDateValidator extends AmendmentDateValidator {
 
   private static final LocalDate EARLIEST_TRANSFER_DATE = LocalDate.of(1995, Month.JANUARY, 1);
 
-  private final DateWrapperUtil dateWrapperUtil;
-
-  public CivilTransferDateValidator(MessageSource messageSource, DateWrapperUtil dateWrapperUtil) {
+  public CivilTransferDateValidator(MessageSource messageSource) {
     super(messageSource);
-    this.dateWrapperUtil = dateWrapperUtil;
   }
 
   @Override
@@ -31,16 +27,15 @@ public class CivilTransferDateValidator extends AmendmentDateValidator {
   @Override
   public void validate(
       ClaimDetails claim, ClaimViewField<?> field, AmendmentForm form, Errors errors) {
-    var transferDate = form.getDateValue(field.name());
+    var transferDate =
+        form.getDateValue(field.toString());
 
-    if (transferDate == null) {
-      return;
-    }
-
-    if (dateWrapperUtil.isFutureDate(transferDate)) {
-      addDateInTheFutureMessage(errors, field);
-    } else if (transferDate.isBefore(EARLIEST_TRANSFER_DATE)) {
-      addDateTooEarlyMessage(errors, field, EARLIEST_TRANSFER_DATE);
+    if (transferDate != null
+        && transferDate.isBefore(EARLIEST_TRANSFER_DATE)) {
+      addDateTooEarlyMessage(
+          errors,
+          field,
+          EARLIEST_TRANSFER_DATE);
     }
   }
 }
