+++ laa-amend-a-claim-ui/src/test/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/CivilTransferDateValidatorTest.java
@@ -1,37 +1,29 @@
 package uk.gov.justice.laa.amend.claim.forms.amendments.validators;
 
+
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.when;
 
 import java.util.Map;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.mockito.Mock;
-import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.validation.BeanPropertyBindingResult;
 import org.springframework.validation.Errors;
 import uk.gov.justice.laa.amend.claim.forms.amendments.AmendmentForm;
 import uk.gov.justice.laa.amend.claim.models.ClaimDetails;
 import uk.gov.justice.laa.amend.claim.resources.MockClaimsFunctions;
 import uk.gov.justice.laa.amend.claim.support.TestMessageSources;
-import uk.gov.justice.laa.amend.claim.utils.DateWrapperUtil;
 import uk.gov.justice.laa.amend.claim.viewmodels.viewfield.CivilClaimDetailsViewField;
 
-@ExtendWith(MockitoExtension.class)
 class CivilTransferDateValidatorTest {
 
   CivilTransferDateValidator validator;
 
-  @Mock DateWrapperUtil dateWrapperUtil;
-
   @BeforeEach
   void beforeEach() {
-    validator = new CivilTransferDateValidator(TestMessageSources.real(), dateWrapperUtil);
+    validator = new CivilTransferDateValidator(TestMessageSources.real());
   }
 
   @Test
@@ -41,8 +33,7 @@ class CivilTransferDateValidatorTest {
 
   @ParameterizedTest
   @MethodSource(
-      "uk.gov.justice.laa.amend.claim.forms.amendments.validators"
-          + ".InvalidDateArgumentsProvider#invalidDateProvider")
+      "uk.gov.justice.laa.amend.claim.forms.amendments.validators.InvalidDateArgumentsProvider#invalidDateProvider")
   void shouldNotAddErrorsForInvalidDates(String day, String month, String year) {
     Map<String, String> input =
         Map.of(
@@ -98,24 +89,6 @@ class CivilTransferDateValidatorTest {
         .isEqualTo("Transfer date");
     assertThat(result.getFieldError("inputs['TRANSFER_DATE']").getArguments()[1])
         .isEqualTo("1 January 1995");
-  }
-
-  @Test
-  void shouldAddErrorsForDateInFuture() {
-    when(dateWrapperUtil.isFutureDate(any())).thenReturn(true);
-    Map<String, String> input =
-        Map.of(
-            "TRANSFER_DATE-day", "2",
-            "TRANSFER_DATE-month", "1",
-            "TRANSFER_DATE-year", "2020");
-
-    Errors result = validate(input);
-
-    assertThat(result.hasFieldErrors()).isTrue();
-    assertThat(result.getFieldError("inputs['TRANSFER_DATE']").getCode())
-        .isEqualTo(AmendmentDateValidator.DATE_CANT_BE_IN_FUTURE_CODE);
-    assertThat(result.getFieldError("inputs['TRANSFER_DATE']").getArguments()[0])
-        .isEqualTo("Transfer date");
   }
 
   private Errors validate(Map<String, String> inputs) {
