+++ laa-amend-a-claim-ui/src/test/java/uk/gov/justice/laa/amend/claim/forms/amendments/validators/CrimeMinRepresentationOrderDateValidatorTest.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.amend.claim.forms.amendments.validators.unique;
+package uk.gov.justice.laa.amend.claim.forms.amendments.validators;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertTrue;
